Update the PCB using the bezier curve spiral coil component. See primary-pcb-coil-bezier.png for a preview of the result.
It requires using the latest Nightly Development Builds from https://kicad-downloads.s3.cern.ch/index.html?prefix=windows/nightly/.
However, as noted in the previous commit, it will be very very slow to load (>15 minutes).
I have a patch for this issue that I will submit.
In the meantime, if anyone is interested to use this please contact me and I will send it to you.